Extract cidr.lang dependent code in //clwb to optional runtime dep
Provide access via EP
