+++ clwb/src/com/google/idea/blaze/clwb/oclang/run/test/BlazeCppTestLocator.java
@@ -16,7 +16,7 @@
 package com.google.idea.blaze.clwb.oclang.run.test;
 
 import com.google.common.collect.ImmutableList;
-import com.google.idea.blaze.clwb.CidrGoogleTestUtilAdapter;
+import com.google.idea.blaze.clwb.oclang.run.CidrGoogleTestUtilAdapter;
 import com.intellij.execution.Location;
 import com.intellij.execution.testframework.sm.runner.SMTestLocator;
 import com.intellij.openapi.project.Project;
